# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SOLOS+ ESPORTZ API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register new member (pending approval)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered, pending approval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
  /auth/forgot-password:
    post:
      operationId: forgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordInput"
      responses:
        "200":
          description: Reset email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Reset password with token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # USERS / PROFILES
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"

    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /announcements/{id}/pin:
    patch:
      operationId: pinAnnouncement
      tags: [announcements]
      summary: Pin/unpin announcement (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"

  # EVENTS
...
  /admin/members/{id}/approve:
    patch:
      operationId: approveMember
      tags: [admin]
      summary: Approve pending member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/members/{id}/reject:
    patch:
      operationId: rejectMember
      tags: [admin]
      summary: Reject pending member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/members/{id}/suspend:
    patch:
      operationId: suspendMember
      tags: [admin]
      summary: Suspend member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/members/{id}/restore:
    patch:
      operationId: restoreMember
      tags: [admin]
      summary: Restore suspended/rejected member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /admin/members/{id}/delete:
    delete:
      operationId: deleteMember
      tags: [admin]
      summary: Delete member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /admin/members/{id}/role:
    patch:
      operationId: updateMemberRole
      tags: [admin]
      summary: Promote/demote member
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleInput"
      responses:
        "200":
          description: Role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # AUTH
    RegisterInput:
      type: object
      required: [username, email, whatsapp, password]
      properties:
        username:
          type: string
          minLength: 2
        email:
          type: string
        whatsapp:
          type: string
        password:
          type: string
          minLength: 8
        tiktok:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        discord:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    ForgotPasswordInput:
      type: object
      required: [email]
      properties:
        email:
          type: string
    ResetPasswordInput:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
        password:
          type: string
          minLength: 8
    AuthResponse:
      type: object
      required: [user]
      properties:
        user:
          $ref: "#/components/schemas/User"

    # USER
    User:
      type: object
      required: [id, username, email, whatsapp, role, status, clanTag, joinedAt]
      properties:
        id:
          type: integer
        username:
          type: string
        clanTag:
          type: string
        email:
          type: string
        whatsapp:
          type: string
        tiktok:
          type: ["string", "null"]
        instagram:
          type: ["string", "null"]
        discord:
          type: ["string", "null"]
        role:
          type: string
          enum: [CLAN_MASTER, ADMIN, TIER1, TIER2, TIER3, NEW_MEMBER]
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED, REJECTED]
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        kills:
          type: integer
        deaths:
          type: integer
        wins:
          type: integer
        losses:
          type: integer
        kdRatio:
          type: number
        clanPoints:
          type: integer
        mvpCount:
          type: integer
        activity:
          type: integer
        tournamentWins:
          type: integer
        scrimWins:
          type: integer
        joinedAt:
          type: string
...
    RoleInput:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [CLAN_MASTER, ADMIN, TIER1, TIER2, TIER3, NEW_MEMBER]

    # ANNOUNCEMENTS
    Announcement:
      type: object
      required: [id, title, content, isPinned, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        isPinned:
          type: boolean
        authorId:
          type: ["integer", "null"]
        authorName:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        isPinned:
          type: boolean
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: ["string", "null"]
        content:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        link:
          type: ["string", "null"]
        isPinned:
          type: ["boolean", "null"]
    PinInput:
      type: object
      required: [isPinned]
      properties:
        isPinned:
          type: boolean

    # EVENTS
...