# lib/api-zod/src/generated/types/announcement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOLOS+ ESPORTZ API
 * OpenAPI spec version: 0.1.0
 */

export interface Announcement {
  id: number;
  title: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  /** @nullable */
  link?: string | null;
  isPinned: boolean;
  /** @nullable */
  authorId?: number | null;
  /** @nullable */
  authorName?: string | null;
  createdAt: string;
  updatedAt?: string;
}