# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SOLOS+ ESPORTZ API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';
import type { UserStatus } from './userStatus';

export interface User {
  id: number;
  username: string;
  clanTag: string;
  email: string;
  whatsapp: string;
  /** @nullable */
  tiktok?: string | null;
  /** @nullable */
  instagram?: string | null;
  /** @nullable */
  discord?: string | null;
  role: UserRole;
  status: UserStatus;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  kills?: number;
  deaths?: number;
  wins?: number;
  losses?: number;
  kdRatio?: number;
  clanPoints?: number;
  mvpCount?: number;
  activity?: number;
  tournamentWins?: number;
  scrimWins?: number;
  joinedAt: string;
}